Medizin Chatbot/requirements.txt
eventlet==0.26.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
Werkzeug==0.16.1
nltk==3.5
Flask-PyMongo==2.3.0
tensorflow-cpu==2.3.0
pusher==3.0.0